physical-ai-book/backend/requirements.txt
# FastAPI and core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
better-auth==0.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Vector Database
qdrant-client==1.7.0

# OpenAI and LangChain
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==24.1.0
flake8==7.0.0
mypy==1.8.0
